// src/generated/prisma/models/User.ts
export type UserModel = runtime.Types.Result.DefaultSelection<Prisma.$UserPayload>
// ...
export type $UserPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "User"
  objects: {
    orders: Prisma.$OrderPayload<ExtArgs>[]
    paymentMethods: Prisma.$PaymentMethodPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    name: string
    email: string
    password: string
    role: $Enums.Role
    country: $Enums.Country
  }, ExtArgs["result"]["user"]>
  composites: {}
}